Documentation for CSSerial API
