+++ Proximity_Sensor_In/Proximity_Sensor_In.ino
@@ -1,36 +1,38 @@
 #include <Wire.h>
 #include "VCNL4010.h"
+#include <CSSerialProtocol.h>
 
-VCNL4010 Sensor;
+VCNL4010 sensor;
 boolean currentTrigger;
 boolean prevTrigger;
 
 int ballCount = 5;
 
 void setup() {
-  Serial.begin(9600);
-  while(!Sensor.begin()) {
-    Serial.println("Proxy Sensor not connected");
-    delay(3000);
+  initSerial(0xB02D);
+  while(!sensor.begin()) {
+    asm volatile ("  jmp 0"); 
   }
-  Sensor.setProximityContinuous(true);
-  Sensor.setProximityHz(250);
-  Sensor.setLEDmA(200);
+  sensor.setProximityContinuous(true);
+  sensor.setProximityHz(250);
+  sensor.setLEDmA(200);
 }
+
+bool ball = false;
 
 void loop() {
   // put your main code here, to run repeatedly:
-  Sensor.setInterrupt(1,false,false,true,false,0,15000);
-  Sensor.getProximity();
-  if(Sensor.getInterrupt() == bit(0)) {
+  sensor.setInterrupt(1,false,false,true,false,0,15000);
+  sensor.getProximity();
+  if(sensor.getInterrupt() == bit(0)) {
     currentTrigger = true;
-    Sensor.clearInterrupt(bit(0));
+    sensor.clearInterrupt(bit(0));
   }
   else{
     currentTrigger = false;
     if(currentTrigger == false && prevTrigger == true) {
       ballCount--;
-      Serial.println("PRXYI increment");
+      inc = true;
     }
   }
   prevTrigger = currentTrigger;
